packages/agent/requirements.txt
langchain==0.1.4
langchain-google-genai==0.0.6
google-generativeai==0.3.2
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.3
streamlit>=1.28.0

